Moves all the implementations of manager components into the hierarchy beneath it. Creates concept of a caching manager & simple manager
