+++ b2/src/b2/manager/documents/document.go
@@ -1,8 +1,8 @@
 package documents
 
 import (
-	"b2/docexmappings"
 	"b2/manager"
+	"b2/manager/docexmappings"
 	"errors"
 	"sync"
 )
